# cmd/api/docs/swagger.yaml
definitions:
...
  models.CategorytOfferResp:
    properties:
      category_id:
        type: integer
      discount_percentage:
        type: integer
      offer_name:
        type: string
    required:
    - category_id
    - discount_percentage
    - offer_name
    type: object
...
  models.ProductOfferResp:
    properties:
      discount_percentage:
        type: integer
      offer_name:
        type: string
      product_id:
        type: integer
    required:
    - discount_percentage
    - offer_name
    - product_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Watch Hive is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Watch Hive
  version: 1.0.0
paths:
...
  /offers:
    post:
      consumes:
      - application/json
      description: Add a new product offer.
      parameters:
      - description: Product offer details in JSON format
        in: body
        name: productOffer
        required: true
        schema:
          $ref: '#/definitions/models.ProductOfferResp'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully added offer
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or constraints not satisfied
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to add offer
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add Product Offer
      tags:
      - Admin Offer Management
  /offers/category:
    get:
      consumes:
      - application/json
      description: Retrieve all category offers.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully got all offers
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or fields provided in the wrong format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve offers
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Category Offer
      tags:
      - Admin Offer Management
    post:
      consumes:
      - application/json
      description: Add a new category offer.
      parameters:
      - description: Category offer details in JSON format
        in: body
        name: categoryOffer
        required: true
        schema:
          $ref: '#/definitions/models.CategorytOfferResp'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully added offer
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or constraints not satisfied
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to add offer
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add Category Offer
      tags:
      - Admin Offer Management
  /offers/category/expire:
    put:
      consumes:
      - application/json
      description: Expire a category offer by providing its ID.
      parameters:
      - description: ID of the category offer to expire
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully made category offer invalid
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or fields provided in the wrong format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to expire category offer
          schema:
            $ref: '#/definitions/response.Response'
      summary: Expire Category Offer
      tags:
      - Admin Offer Management
  /offers/product:
    get:
      consumes:
      - application/json
      description: Retrieve all product offers.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully got all offers
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or fields provided in the wrong format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve offers
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Product Offer
      tags:
      - Admin Offer Management
  /offers/product/expire:
    put:
      consumes:
      - application/json
      description: Expire a product offer by providing its ID.
      parameters:
      - description: ID of the product offer to expire
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully made product offer invalid
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request format or fields provided in the wrong format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to expire product offer
          schema:
            $ref: '#/definitions/response.Response'
      summary: Expire Product Offer
      tags:
      - Admin Offer Management
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"